### 2024-09-09 #258 @editorialbot
[PRE REVIEW]: Learning from a multimodal approach with sonoTraining platform
**Submitting author:** @NatashaBertaina (<a href="http://orcid.org/0009-0001-2722-4675">Natasha María Monserrat Bertaina Lucero</a>)
**Repository:** https://github.com/NatashaBertaina/webTraining
**Branch with paper.md** (empty if default branch): jose-paper
**Version:** v1.0
