+++ userApp/views.py
@@ -1,8 +1,6 @@
 from django.shortcuts import render, redirect
-from django.views import generic, View
-from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
-
-from django.contrib.auth.forms import UserCreationForm, AuthenticationForm
+from django.views import View
+from django.contrib.auth.forms import AuthenticationForm
 from django.contrib.auth import login, logout, authenticate
 from django.contrib.auth.models import User
 from django.contrib.auth.decorators import login_required
@@ -13,12 +11,20 @@ from django.db import transaction
 from django.contrib.auth.mixins import LoginRequiredMixin  #LoginRequiredMixin se utiliza como un mixin para requerir que un usuario esté autenticado antes de acceder a una vista específica.
 from django.contrib.auth.views import PasswordChangeView, PasswordResetDoneView
 from django.urls import reverse_lazy
-from django.contrib.auth.forms import PasswordChangeForm
 from trainingApp.models import Group
-
+from django.core.mail import send_mail
+from django.contrib.sites.shortcuts import get_current_site
+from django.template.loader import render_to_string
+from django.utils.http import urlsafe_base64_encode
+from django.utils.encoding import force_bytes, force_str
+from django.contrib.auth.tokens import default_token_generator
+from django.utils.http import urlsafe_base64_decode
+from dotenv import load_dotenv
+import os
+# Cargar las variables de entorno desde el archivo .env
+load_dotenv()
 
 ##Logica de users 
-
 def home(request):
     return render(request, 'home.html')
 
@@ -30,23 +36,59 @@ def signup(request):
     else:
         form = SignupForm(request.POST)
         if form.is_valid():
-            user = form.save()
+            user = form.save(commit=False)
+            user.is_active = False  # Desactiva la cuenta hasta que se confirme el email
             # Crear el objeto Trainee y se relaciona con el usuario y al grupo por defecto
-            group_defect = Group.objects.get(name_group = 'Group Defect')
-            Trainee.objects.create(
-                user=user,
-                age=form.cleaned_data['age'],
-                educationalLevel=form.cleaned_data['educationalLevel'],
-                occupation=form.cleaned_data['occupation'],
-                group =  group_defect
-            )
-            messages.success(request,"Your account is crated successfully")
-            login(request, user)
+            group_defect, created = Group.objects.get_or_create(name_group='Group Defect')
+            trainee = Trainee(
+                    user=user,
+                    age=form.cleaned_data['age'],
+                    educationalLevel=form.cleaned_data['educationalLevel'],
+                    occupation=form.cleaned_data['occupation'],
+                    group=group_defect
+                )
+            user.save()
+            trainee.save()
+            # Enviar email de confirmación
+            try:
+                current_site = get_current_site(request)
+                mail_subject = 'Activate your account.'
+                mail_transmitter = os.getenv('EMAIL_HOST_USER')
+                message = render_to_string('userApp/acc_active_email.html', {
+                    'user': user,
+                    'domain': current_site.domain,
+                    'uid': urlsafe_base64_encode(force_bytes(user.pk)),  # uid se utiliza para identificar de manera única al usuario en el enlace de activación sin exponer directamente el ID del usuario en la URL.
+                    'token': default_token_generator.make_token(user),  # Crea un token único basado en el usuario y otros factores
+                })
+                send_mail(mail_subject, message, mail_transmitter, [user.email]) 
+            except Exception as e:
+                messages.error(request, "There was an error sending the confirmation email. Please try again later.")
+                return redirect('home')
             
-            return redirect('trainingApp:training_List')
[…]
