+++ .github/workflows/paper.yml
@@ -6,7 +6,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v3
+        uses: actions/checkout@v4
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -14,7 +14,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
           # Bibliography file, if existing
-          bibtex: paper/paper.bib
+          #bibtex: paper/paper.bib
       - name: Upload
         uses: actions/upload-artifact@v1
         with:
