+++ .github/workflows/paper.yml
@@ -6,7 +6,7 @@ jobs:
     name: Paper Draft
     steps:
       - name: Checkout
-        uses: actions/checkout@v4
+        uses: actions/checkout@v3
       - name: Build draft PDF
         uses: openjournals/openjournals-draft-action@master
         with:
@@ -14,9 +14,9 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: paper/paper.md
           # Bibliography file, if existing
-          #bibtex: paper/paper.bib
+          bibtex: paper/paper.bib
       - name: Upload
-        uses: actions/upload-artifact@v2
+        uses: actions/upload-artifact@v1
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
