### 2024-10-02 e4424db
Update paper.yml
