+++ trainingApp/migrations/0009_group_training_gruops.py
@@ -1,26 +1,0 @@
-# Generated by Django 5.0.1 on 2024-03-31 18:00
-
-from django.db import migrations, models
-
-
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('trainingApp', '0008_alter_blockanswer_state_block'),
-    ]
-
-    operations = [
-        migrations.CreateModel(
-            name='Group',
-            fields=[
-                ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
-                ('name_group', models.CharField(max_length=100)),
-                ('description', models.CharField(max_length=300)),
-            ],
-        ),
-        migrations.AddField(
-            model_name='training',
-            name='gruops',
-            field=models.ManyToManyField(to='trainingApp.group'),
-        ),
-    ]
