+++ userApp/migrations/0001_initial.py
@@ -1,27 +1,0 @@
-# Generated by Django 5.0.1 on 2024-02-28 22:03
-
-import django.db.models.deletion
-from django.conf import settings
-from django.db import migrations, models
-
-
-class Migration(migrations.Migration):
-
-    initial = True
-
-    dependencies = [
-        migrations.swappable_dependency(settings.AUTH_USER_MODEL),
-    ]
-
-    operations = [
-        migrations.CreateModel(
-            name='Trainee',
-            fields=[
-                ('id', models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
-                ('age', models.IntegerField(default=18)),
-                ('educationalLevel', models.CharField(choices=[('Primary', 'Primary School'), ('High School', 'High School'), ('Collage', 'Collage'), ('Doctoral Student', 'Doctoral'), ('Ph.D', 'Graduate')], default='High School', max_length=20)),
-                ('occupation', models.CharField(max_length=50)),
-                ('user', models.OneToOneField(on_delete=django.db.models.deletion.CASCADE, to=settings.AUTH_USER_MODEL)),
-            ],
-        ),
-    ]
