+++ userApp/migrations/0004_alter_trainee_educationallevel.py
@@ -1,18 +1,0 @@
-# Generated by Django 5.0.1 on 2024-04-12 18:05
-
-from django.db import migrations, models
-
-
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('userApp', '0003_alter_trainee_group'),
-    ]
-
-    operations = [
-        migrations.AlterField(
-            model_name='trainee',
-            name='educationalLevel',
-            field=models.CharField(choices=[('Primario', 'Primary School'), ('Secundario', 'High School'), ('Terciario', 'Associate'), ('Universitario', 'University'), ('Estudiante de doctorado', 'Doctoral Student'), ('Magister', 'Magister'), ('Doctor/Doctora', 'Doctor')], default='Secundario', max_length=50),
-        ),
-    ]
