+++ trainingApp/migrations/0013_block_description_en_block_description_es_and_more.py
@@ -1,33 +1,0 @@
-# Generated by Django 5.0.4 on 2024-08-26 16:40
-
-from django.db import migrations, models
-
-
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('trainingApp', '0012_choice_choice_en_choice_choice_es_and_more'),
-    ]
-
-    operations = [
-        migrations.AddField(
-            model_name='block',
-            name='description_en',
-            field=models.CharField(max_length=500, null=True),
-        ),
-        migrations.AddField(
-            model_name='block',
-            name='description_es',
-            field=models.CharField(max_length=500, null=True),
-        ),
-        migrations.AddField(
-            model_name='block',
-            name='name_block_en',
-            field=models.CharField(max_length=100, null=True),
-        ),
-        migrations.AddField(
-            model_name='block',
-            name='name_block_es',
-            field=models.CharField(max_length=100, null=True),
-        ),
-    ]
