+++ trainingApp/migrations/0010_rename_gruops_training_groups.py
@@ -1,18 +1,0 @@
-# Generated by Django 5.0.1 on 2024-03-31 18:19
-
-from django.db import migrations
-
-
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('trainingApp', '0009_group_training_gruops'),
-    ]
-
-    operations = [
-        migrations.RenameField(
-            model_name='training',
-            old_name='gruops',
-            new_name='groups',
-        ),
-    ]
