+++ trainingApp/migrations/0004_blockdeploy_state_block_deploy.py
@@ -1,18 +1,0 @@
-# Generated by Django 5.0.1 on 2024-03-15 15:12
-
-from django.db import migrations, models
-
-
-class Migration(migrations.Migration):
-
-    dependencies = [
-        ('trainingApp', '0003_remove_deploy_deploy_type_remove_deploy_training_and_more'),
-    ]
-
-    operations = [
-        migrations.AddField(
-            model_name='blockdeploy',
-            name='state_block_deploy',
-            field=models.CharField(choices=[('in progress', 'In Progress'), ('Complated', 'Complated')], default='in progress', max_length=20),
-        ),
-    ]
