### 2025-05-30 2084261
refactor: eliminar función de confirmacion email
