+++ trainingApp/migrations/0004_block_description_en_block_description_es_and_more.py
@@ -1,0 +1,63 @@
+# Generated by Django 5.0.4 on 2025-10-31 22:58
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('trainingApp', '0003_initial'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='block',
+            name='description_en',
+            field=models.CharField(max_length=500, null=True),
+        ),
+        migrations.AddField(
+            model_name='block',
+            name='description_es',
+            field=models.CharField(max_length=500, null=True),
+        ),
+        migrations.AddField(
+            model_name='block',
+            name='name_block_en',
+            field=models.CharField(max_length=100, null=True),
+        ),
+        migrations.AddField(
+            model_name='block',
+            name='name_block_es',
+            field=models.CharField(max_length=100, null=True),
+        ),
+        migrations.AddField(
+            model_name='choice',
+            name='choice_en',
+            field=models.CharField(max_length=100, null=True),
+        ),
+        migrations.AddField(
+            model_name='choice',
+            name='choice_es',
+            field=models.CharField(max_length=100, null=True),
+        ),
+        migrations.AddField(
+            model_name='deploy',
+            name='question_en',
+            field=models.CharField(max_length=100, null=True),
+        ),
+        migrations.AddField(
+            model_name='deploy',
+            name='question_es',
+            field=models.CharField(max_length=100, null=True),
+        ),
+        migrations.AddField(
+            model_name='training',
+            name='name_training_en',
+            field=models.CharField(max_length=200, null=True),
+        ),
+        migrations.AddField(
+            model_name='training',
+            name='name_training_es',
+            field=models.CharField(max_length=200, null=True),
+        ),
+    ]
