+++ requirements.txt
@@ -22,5 +22,5 @@ virtualenvwrapper-win==1.2.7
 whitenoise==6.6.0
 xlrd==2.0.1
 xlwt==1.3.0
-django-modeltranslation== 0.19.7
-python-dotenv== 1.0.1
+django-modeltranslation==0.19.7
+python-dotenv==1.0.1
