### 2025-11-04 35d57df
scripts training mariabd
