+++ requirements.txt
@@ -7,6 +7,7 @@ django-import-export==3.3.8
 et-xmlfile==1.1.0
 filelock==3.12.4
 MarkupPy==1.14
+mysqlclient==2.2.7
 odfpy==1.4.1
 openpyxl==3.1.2
 pillow==10.3.0
