+++ userApp/urls.py
@@ -23,5 +23,5 @@ urlpatterns = [
     ), name='password_reset'),
     path('password_reset/done/', auth_views.PasswordResetDoneView.as_view(template_name="userApp/password_reset_done.html"), name='password_reset_done'),
     path('reset/<uidb64>/<token>/', CustomPasswordResetConfirmView.as_view(template_name="userApp/passwordChange.html"), name='password_reset_confirm'),
-    path('reset/done/', auth_views.PasswordResetCompleteView.as_view(template_name="userApp/passwordresetDone.html"), name='password_reset_complete'),
+    path('reset/done/', auth_views.PasswordResetCompleteView.as_view(template_name="userApp/passwordResetDone.html"), name='password_reset_complete'),
 ]
