### 2025-11-21 43acf32
Admin panel
