+++ webTraining/urls.py
@@ -15,6 +15,7 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.contrib import admin
+from . import custom_admin  # apply admin UI tweaks (reorder models in trainingApp)
 from django.urls import path, include
 from django.conf import settings
 from django.conf.urls.static import static
