+++ trainingApp/admin.py
@@ -4,6 +4,30 @@ from import_export.admin import ExportActionMixin
 from import_export import resources  # Importa resources desde import_export
 from import_export.fields import Field
 from modeltranslation.admin import TranslationAdmin
+from django.utils.html import format_html
+from django.urls import reverse
+from django.http import HttpResponseRedirect
+
+
+# Mixin para controlar la paginación desde un parámetro GET (?per_page=10|50|100)
+class PaginationMixin:
+    per_page_choices = (10,)
+    default_list_per_page = 20
+
+    def changelist_view(self, request, extra_context=None):
+        session_key = f'admin_per_page_{self.model._meta.app_label}_{self.model._meta.model_name}'
+        per_page_i = request.session.get(session_key)
+        if per_page_i in self.per_page_choices:
+            self.list_per_page = per_page_i
+        else:
+            self.list_per_page = self.default_list_per_page
+
+        if extra_context is None:
+            extra_context = {}
+        extra_context['per_page_choices'] = self.per_page_choices
+        extra_context['current_per_page'] = getattr(self, 'list_per_page', self.default_list_per_page)
+        extra_context['request'] = request
+        return super().changelist_view(request, extra_context=extra_context)
 
 #Clase que define cómo se importarán/exportarán los objetos DeployAnswer.
 class DeployAnswerResource(resources.ModelResource): 
@@ -35,8 +59,22 @@ class DeployResource(resources.ModelResource):
         
 
 #Clases que heredan de admin.ModelAdmin gestionan el estilo en el admin
-class TrainingAdmin(TranslationAdmin,admin.ModelAdmin):
-    list_display = ('name_training',)
+class TrainingAdmin(PaginationMixin, TranslationAdmin, admin.ModelAdmin):
+    list_display = ('id','name_training','view_blocks','difficulty','state_training','estimatedDuration','attempts_allowed','pub_date')
+    list_display_links = ('name_training',)
+    list_filter = ('difficulty','state_training','groups')
+    search_fields = ('name_training',)
+    ordering = ('-pub_date',)
+
+    def view_blocks(self,obj):
+        """Link to the Block changelist filtered by this training."""
+        try:
+            count = obj.block_set.count()
+        except Exception:
+            count = 0
+        url = reverse('admin:trainingApp_block_changelist') + f'?training__id__exact={obj.id}'
+        return format_html('<a href="{}">Blocks ({})</a>', url, count)
+    view_blocks.short_description = 'Blocks'
     
     class Media:
         js = (
@@ -51,8 +89,13 @@ class TrainingAdmin(TranslationAdmin,admin.ModelAdmin):
             ),
         }
         
-class BlockAdmin(TranslationAdmin,admin.ModelAdmin):
-    list_display = ('name_block', 'description',) 
+class BlockAdmin(PaginationMixin, TranslationAdmin, admin.ModelAdmin):
+    list_display = ('id','name_block','view_deploys', 'training', 'state_block','estimed_duration_block',)
+    list_display_links = ('name_block',)
+    list_filter = ('state_block','training')
+    search_fields = ('name_block','description')
+    list_select_related = ('training',)
+    autocomplete_fields = ('training',)
     
     class Media:
         js = (
@@ -66,9 +109,37 @@ class BlockAdmin(TranslationAdmin,admin.ModelAdmin):
                 'https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css',  # Nuevo estilo añadido
             ),
         }
+
+    def view_deploys(self,obj):
[…]
