### 2026-03-19 #258 @bleds22e
Hello @NatashaBertaina and @johicasado!
First of all, let me offer our sincere apologies for the long delay since your initial submission. Please let me know if you're still interested in having this paper reviewed at JOSE. If you are, I will work on getting the review process started ASAP.
