### 2026-03-26 #258 @bleds22e
@johicasado Sounds good! I will assign myself as editor and start asking for reviewers; thank you for the suggestions above.

Is there anything in the paper you would like to update before we start the review process?
