### 2026-03-26 #258 @djmannion
@bleds22e, yes I would be happy to review this paper.
