### 2026-04-10 #258 @johicasado
@bleds22e I will change some details about affiliations only. About the possible reviewers, I'm not sure. I checked the list from JOSE, maybe

- Ariel Rokem | arokem@gmail.com | arokem
- Christopher McComb | mccomb@psu.edu | cmccomb
- Nathan Moore | ntmoore@gmail.com | ntmoore
- Stefania Druga | sdruga@mit.edu | stefania11
