### 2026-04-17 #328 @editorialbot
[REVIEW]: Learning from a multimodal approach with sonoTraining platform
**Submitting author:** @NatashaBertaina (<a href="http://orcid.org/0009-0001-2722-4675">Natasha María Monserrat Bertaina Lucero</a>)
**Repository:** https://github.com/NatashaBertaina/webTraining
**Branch with paper.md** (empty if default branch): jose-paper
**Version:** v1.0
**Paper kind:** software
