### 2026-04-22 NatashaBertaina/webTraining#17 @cmccomb
Create a v1.0 release
Hi team! I am currently conducting the [JOSE review](https://github.com/openjournals/jose-reviews/issues/328) for this repository.

To satisfy the journal's requirements regarding software versioning, could you please create a formal **GitHub Release** for the repository? Please tag the release as `v1.0` to match the version specified in the review issue.

Thank you for your help in moving this review forward!
