### 2026-04-22 NatashaBertaina/webTraining#18 @cmccomb
Address missing paper sections
Hi team! I am currently conducting the [JOSE review](https://github.com/openjournals/jose-reviews/issues/328) for this repository.

It looks like your paper is currently missing four required sections (State of the field, Software design, Research impact statement, and AI usage disclosure).

Please let me know once these updates have been pushed to the repository so I can continue with the review. Thanks!
