### 2026-04-22 NatashaBertaina/webTraining#19 @cmccomb
Update Trayford citation to use DOI
Hi authors! I am currently conducting the [JOSE review](https://github.com/openjournals/jose-reviews/issues/328) for this repository.

In the current `paper.md`, I noticed the citation for the Trayford et al. paper points to the arXiv preprint version. To ensure the paper points to the most permanent and citable record, please update this reference to use the formal DOI for the published version: [https://doi.org/10.1093/rasti/rzad021](https://doi.org/10.1093/rasti/rzad021).

Thanks!
