### 2026-04-22 NatashaBertaina/webTraining#20 @cmccomb
Implement functional test suite
Hi authors! I am currently conducting the [JOSE review](https://github.com/openjournals/jose-reviews/issues/328) for this repository.

While some test files exist, they currently contain only the default Django boilerplate code and no implemented test cases. To satisfy the requirements regarding software quality and reproducibility, please implement a test suite that verifies the core functionality of the application. Of course, please let me know if I missed the tests in my review. Thanks!
