### 2026-04-22 NatashaBertaina/webTraining#21 @cmccomb
Provide comprehensive installation and setup documentation
Hi authors! I am currently conducting the [JOSE review](https://github.com/openjournals/jose-reviews/issues/328) for this repository. To ensure the software is reproducible, please update the README to include clear installation instructions. Currently, the repository lacks documentation on how to install everything and get up and running. Some of this can be reverse engineered from GitHub Actions, but a user-facing walkthrough would be valuable. Thanks!
