### 2026-04-26 NatashaBertaina/webTraining#22 @cmccomb
Add clear community and support guidelines
Hi authors! I am currently conducting the [JOSE review](https://github.com/openjournals/jose-reviews/issues/328) for this repository.

The JOSE review checklist asks whether the repository includes clear guidelines for third parties wishing to: 1) contribute to the software, 2) report issues or problems, and 3) seek support. The README currently includes a brief contribution/contact note and mentions opening GitHub issues, but it does not yet provide clear guidance for all three of these paths.

Could you please add community/support guidelines, either in the README or in dedicated files such as `CONTRIBUTING.md` and/or `SUPPORT.md`? I consider this a blocker for checking off the Community guidelines item.

Thanks!
