### 2026-04-26 NatashaBertaina/webTraining#23 @cmccomb
Document an end-to-end usage walkthrough
Hi authors! I am currently conducting the [JOSE review](https://github.com/openjournals/jose-reviews/issues/328) for this repository.

In addition to installation instructions, the repository needs an end-to-end usage walkthrough showing how a new user or reviewer can exercise the main functionality of the software. For example, this could include how to create an account or admin user, load sample training data, access a course, complete a training block, and verify that responses/results are recorded.

Without this walkthrough, I cannot confidently check off the Example usage or Functionality documentation items in the JOSE review checklist. If manual verification steps are included, this would also help satisfy the Tests checklist item.

Thanks!
